+++ client/src/components/Register.js
@@ -1,0 +1,21 @@
+import React from "react";
+import { Login } from "./Register/Login";
+import { Signup } from "./Register/Signup";
+
+export default function Register(){
+    const [currentForm, setCurrentForm] = useState("login");
+
+  const toggleForm = (formName) => {
+    setCurrentForm(formName);
+  };
+  return (
+    <div>
+      {currentForm === "login" ? (
+        <Login onFormSwitch={toggleForm} />
+      ) : (
+        <Signup onFormSwitch={toggleForm} />
+      )}
+    </div>
+  );
+
+}
